requirements.txt
asn1crypto
certifi
cffi
cryptography
fpdf
idna
pdfkit
Pillow
pycparser
pyOpenSSL
PySocks
selenium
six
urllib3
win-inet-pton
wincertstore
webdriver-manager

